C item

+++ Assets/02. Scrpits/Enemy/EnemyS/Boss/Stage3Boss/Stage3Boss.cs
@@ -51,7 +51,7 @@ public class Stage3Boss : BossBase
         fsm.Tick();
     }
 
-    public override void TakeDamage(float amount)
+    public override void TakeDamage(int amount)
     {
         Stat.curHp -= amount;
         if (Stat.curHp < (Stat.totalHp * 0.5f) && !isSpecial2On)
